Refactor Auth file and directory
