+++ App/Navigation/routes.js
@@ -1,6 +1,6 @@
 import Home from '~/Scene/Home';
-import Login from '~/Scene/Auth/Login';
-import ForgotPassword from '~/Scene/Auth/ForgotPassword';
+import Login from '~/Scene/Auth/Login/';
+import ForgotPassword from '~/Scene/Auth/ForgotPassword/';
 import NewsFeed from '~/Scene/NewsFeed';
 import Notification from '~/Scene/Notification';
 import People from '~/Scene/People';
