+++ App/Scene/Auth/ForgotPassword/Form.js
@@ -1,0 +1,77 @@
+import React, { PureComponent } from 'react';
+import PropTypes from 'prop-types';
+import { Field, reduxForm } from 'redux-form';
+import { required, email } from '~/Lib/validate';
+
+import { KeyboardAvoidingView, Image, View } from 'react-native';
+import { Images } from '~/Theme';
+import FormInput from '~/Component/FormInput';
+import TouchableView from '~/Component/TouchableView';
+import Text from '~./Component/Text';
+import styles from '../styles';
+
+const submit = values => {
+  console.log('submitting form', values);
+};
+
+class ForgotPasswordForm extends PureComponent {
+  static defaultProps = {
+    error: null,
+  };
+
+  static propTypes = {
+    handleSubmit: PropTypes.func,
+    reset: PropTypes.func.isRequired,
+    pristine: PropTypes.bool.isRequired,
+    invalid: PropTypes.bool.isRequired,
+    error: PropTypes.string,
+  };
+
+  render() {
+    const { handleSubmit } = this.props;
+    return (
+      <KeyboardAvoidingView
+        onSubmit={handleSubmit}
+        style={styles.container}
+        behavior={'position'}
+      >
+        <View style={styles.imageContainer}>
+          <Image style={styles.image} source={Images.imgDefault150} />
+          <Text bold style={styles.title}>
+            Forgot Your Password?
+          </Text>
+          <Text light style={styles.description}>
+            Enter your email below to reset your password
+          </Text>
+        </View>
+        <View style={styles.formContainer}>
+          <Field
+            name="email"
+            type="email"
+            component={FormInput}
+            validate={[required, email]}
+            placeholder="Type your Email"
+            underlineColorAndroid={'transparent'}
+            keyboardType={'email-address'}
+          />
+        </View>
+        <View style={styles.buttonContainer}>
+          <TouchableView
+            onPress={handleSubmit(submit)}
+            style={styles.submitButton}
+          >
+            <Text bold style={styles.buttonText}>
+              SEND
+            </Text>
+          </TouchableView>
+        </View>
+      </KeyboardAvoidingView>
+    );
+  }
+}
+
+ForgotPasswordForm = reduxForm({
+  form: 'forgotPassword',
+})(ForgotPasswordForm);
+
+export default ForgotPasswordForm;
